Fix SCRPAD_ADDR truncation and scratchpad initialization
- CRITICAL: SCRPAD_ADDR was (unsigned long) = uint32_t, truncating 64-bit
  scratchpad pointer. Changed to (unsigned long long) for correct 64-bit.
- Zero scratchpad every frame in game_tick (collision stubs don't init it)
- Fix HZD_PointCheck stub: clear scratchpad MAX_FLOOR field
- Game now reaches s00a gameplay (next: sound system crash in SePlay)

+++ port/psx/psxdefs.h
@@ -13,7 +13,7 @@ typedef long (*openevent_cb_t)();
 
 extern char port_scratchpad[1024];
 
-#define SCRPAD_ADDR     ((unsigned long)(port_scratchpad))
+#define SCRPAD_ADDR     ((unsigned long long)(port_scratchpad))
 #define SCRPAD_SIZE     0x400
 
 #define getScratchAddr2(type, offset) ((type *)(port_scratchpad + (offset)))
